refactor(migration-phase-2): use hard-coded value for `dstWant`
`dstWant` is supposed to be set to `type(uin256).max`, otherwise it
might leave room for a flash loan attack that would prevent the
migration from happening.

The scenario was suggested by ChainSecurity:

> 1. Flashloan gem
> 2. Sell gem to the old PSM inflating the ink balance
> 3. Migrate and move maximum (dropping the ink to the pre-manipulation state)
> 4. Buy all the migrated gem from PSM lite and repay the flashloan

While that is actually possible for a large range of values for
`dstWant`, in practice the attacker would be limited by `srcGap` after
migration.  Still, we could end up in a scenario where there is `srcGap`
less gems in `dstPsm`.

However, when `dstWant` is large enough, any artificial inflation of
`srcInk` would be migrated to `dstPsm`.

For that reason, it is better to just hard-code a value and remove it
from the configuration.